requirements.txt
Django  
djangorestframework
pandas 
numpy
scikit-learn
statsmodels
matplotlib
seaborn
plotly
reportlab
psycopg2-binary
django-crispy-forms
crispy-bootstrap5
pmdarima
tensorflow
keras 
django-extensions 
django-two-factor-auth  